social_bmp_backend/requirements.txt
bson
Django
djongo
djangorestframework<4
py
pytest
pytest-django
